+++ server/test/system/market-making/execution/sandbox-fill-resolution.system.spec.ts
@@ -5,10 +5,10 @@ import {
   buildSandboxClientOrderId,
   getSandboxIntegrationSkipReason,
   SandboxExchangeHelper,
-} from '../../../../test/helpers/sandbox-exchange.helper';
-import { ExchangeOrderMapping } from '../../../common/entities/market-making/exchange-order-mapping.entity';
-import { ExchangeOrderMappingService } from './exchange-order-mapping.service';
-import { FillRoutingService } from './fill-routing.service';
+} from '../../helpers/sandbox-exchange.helper';
+import { ExchangeOrderMapping } from '../../../../src/common/entities/market-making/exchange-order-mapping.entity';
+import { ExchangeOrderMappingService } from '../../../../src/modules/market-making/execution/exchange-order-mapping.service';
+import { FillRoutingService } from '../../../../src/modules/market-making/execution/fill-routing.service';
 
 const skipReason = getSandboxIntegrationSkipReason();